requirements.txt
tensorflow==2.19.0
numpy
scikit-learn
matplotlib
streamlit 
Pillow